requirements.txt
aiosqlite
aiogram==3.4.1
python-dotenv
